// angular/src/app/interfaces/company-profile-api-response-interface.ts
export interface CompanyProfileApiResponseInterface {
  address: string;
  city: string;
  date_created: string;
  date_updated: string;
  email: string;
  id: number;
  logoUrl: string;
  name: string;
  state: string;
  street: string;
  user_id: number;
  websiteUrl: string;
  zipCode: string;
  loaded: boolean;
}